pull out kubernetes and log
