+++ batch/batch/server/kubernetes.py
@@ -1,0 +1,9 @@
+import os
+
+import kubernetes as kube
+
+if 'BATCH_USE_KUBE_CONFIG' in os.environ:
+    kube.config.load_kube_config()
+else:
+    kube.config.load_incluster_config()
+v1 = kube.client.CoreV1Api()
