+++ batch/batch/server/log.py
@@ -1,0 +1,26 @@
+import logging
+
+
+def make_logger():
+    fmt = logging.Formatter(
+        # NB: no space after levename because WARNING is so long
+        '%(levelname)s\t| %(asctime)s \t| %(filename)s \t| %(funcName)s:%(lineno)d | '
+        '%(message)s')
+
+    file_handler = logging.FileHandler('batch.log')
+    file_handler.setLevel(logging.INFO)
+    file_handler.setFormatter(fmt)
+
+    stream_handler = logging.StreamHandler()
+    stream_handler.setLevel(logging.INFO)
+    stream_handler.setFormatter(fmt)
+
+    log = logging.getLogger('batch')
+    log.setLevel(logging.INFO)
+
+    logging.basicConfig(handlers=[file_handler, stream_handler], level=logging.INFO)
+
+    return log
+
+
+log = make_logger()
